move BucketQueue from routing.h to bucket_queue.h
